+++ test/observables.js
@@ -932,7 +932,6 @@ test('issue 50', function(t) {
         t.deepEqual(stripSpyOutput(events), [
 			'auto',
 			'calc c',
-			{ name: 'anonymous transaction', spyReportStart: true, target: undefined, type: 'transaction' },
 			'transstart',
 			{ name: 'a', newValue: false, oldValue: true, spyReportStart: true, type: 'update' }, { spyReportEnd: true },
 			{ name: 'b', newValue: true, oldValue: false, spyReportStart: true, type: 'update' }, { spyReportEnd: true },
@@ -941,7 +940,6 @@ test('issue 50', function(t) {
 			'auto',
       { target: { a: false, b: true }, type: 'compute' },
       'calc c',
-			{ spyReportEnd: true },
 			{ spyReportEnd: true },
 			'transpostend'
         ]);
@@ -986,14 +984,12 @@ test('verify transaction events', function(t) {
 	t.deepEqual(stripSpyOutput(events), [
 		'auto',
 		'calc c',
-		{ name: 'anonymous transaction', spyReportStart: true, target: undefined, type: 'transaction' },
 		'transstart',
 		{ name: 'b', newValue: 2, oldValue: 1, spyReportStart: true, type: 'update' }, { spyReportEnd: true },
 		'transpreend', { target: { b: 2 }, type: 'compute' },
 		'calc c',
 		{ spyReportStart: true, type: 'reaction' },
 		'auto',
-		{ spyReportEnd: true },
 		{ spyReportEnd: true },
 		'transpostend'
     ]);
@@ -1073,19 +1069,15 @@ test('delay autorun until end of transaction', function(t) {
     events.push("post trans3");
 
     t.deepEqual(stripSpyOutput(events), [
-		{ name: 'anonymous transaction', spyReportStart: true, target: undefined, type: 'transaction' },
-			{ name: 'anonymous transaction', spyReportStart: true, target: undefined, type: 'transaction' },
-				{ name: 'a', newValue: 3, oldValue: 2, spyReportStart: true, type: 'update' }, { spyReportEnd: true },
-				{ name: 'a', newValue: 4, oldValue: 3, spyReportStart: true, type: 'update' }, { spyReportEnd: true },
-				'end1',
-			{ spyReportEnd: true },
-			{ name: 'a', newValue: 5, oldValue: 4, spyReportStart: true, type: 'update' }, { spyReportEnd: true },
-			'end2',
-			{ spyReportStart: true, type: 'reaction' },
-				'auto',
-				{ target: { a: 3 }, type: 'compute' },
-				'calc y',
-			{ spyReportEnd: true },
+		{ name: 'a', newValue: 3, oldValue: 2, spyReportStart: true, type: 'update' }, { spyReportEnd: true },
+		{ name: 'a', newValue: 4, oldValue: 3, spyReportStart: true, type: 'update' }, { spyReportEnd: true },
+		'end1',
+		{ name: 'a', newValue: 5, oldValue: 4, spyReportStart: true, type: 'update' }, { spyReportEnd: true },
+		'end2',
+		{ spyReportStart: true, type: 'reaction' },
+			'auto',
+			{ target: { a: 3 }, type: 'compute' },
+			'calc y',
 		{ spyReportEnd: true },
 		'post trans1',
 		{ name: 'a', newValue: 6, oldValue: 5, spyReportStart: true, type: 'update' },